Move functions to other file
